+++ src/check_jsonschema/parsers/__init__.py
@@ -1,12 +1,12 @@
 from __future__ import annotations
 
 import json
+import pathlib
 import typing as t
 
+import ruamel.yaml
 from identify import identify
 
-from ...transforms import Transform
-from ..errors import BadFileTypeError
 from . import json5, toml, yaml
 
 DEFAULT_LOAD_FUNC_BY_TAG: dict[str, t.Callable[[t.BinaryIO], t.Any]] = {
@@ -22,14 +22,31 @@ MISSING_SUPPORT_MESSAGES: dict[str, str] = {
 }
 
 
+class BadFileTypeError(ValueError):
+    pass
+
+
 class ParserSet:
-    def __init__(self, transform: Transform):
+    def __init__(
+        self,
+        *,
+        modify_yaml_implementation: t.Callable[[ruamel.yaml.YAML], ruamel.yaml.YAML]
+        | None = None,
+        supported_formats: t.Sequence[str, ...] | None = None,
+    ) -> None:
         yaml_impl = yaml.construct_yaml_implementation()
-        transform.modify_yaml_implementation(yaml_impl)
-        self._by_tag = {
+        if modify_yaml_implementation:
+            modify_yaml_implementation(yaml_impl)
+        base_by_tag = {
             "yaml": yaml.impl2loader(yaml_impl),
             **DEFAULT_LOAD_FUNC_BY_TAG,
         }
+        if supported_formats is None:
+            self._by_tag = base_by_tag
+        else:
+            self._by_tag = {
+                k: v for k, v in base_by_tag.items() if k in supported_formats
+            }
 
     def get(
         self, filename: str, default_ft: str | None
@@ -44,35 +61,21 @@ class ParserSet:
         for tag in tags:
             if tag in MISSING_SUPPORT_MESSAGES:
                 raise BadFileTypeError(
-                    f"cannot check {filename} because support is missing for {tag}\n"
+                    f"cannot parse {filename} because support is missing for {tag}\n"
                     + MISSING_SUPPORT_MESSAGES[tag]
                 )
         raise BadFileTypeError(
-            f"cannot check {filename} as it is not one of the supported filetypes: "
+            f"cannot parse {filename} as it is not one of the supported filetypes: "
             + ",".join(self._by_tag.keys())
         )
 
-
-class InstanceLoader:
-    def __init__(
+    def parse_file(
         self,
-        filenames: t.Sequence[str],
-        default_filetype: str | None = None,
-        data_transform: Transform | None = None,
-    ) -> None:
-        self._filenames = filenames
-        self._default_ft = default_filetype.lower() if default_filetype else None
-        self._data_transform = (
-            data_transform if data_transform is not None else Transform()
+        path: str | pathlib.Path,
+        default_ft: str | None,
+    ):
+        loadfunc = self.get(
+            str(path) if isinstance(path, pathlib.Path) else path, default_ft
         )
-
-        self._parsers = ParserSet(self._data_transform)
-
-    def iter_files(self) -> t.Iterator[tuple[str, t.Any]]:
-        for fn in self._filenames:
-            loadfunc = self._parsers.get(fn, self._default_ft)
-
-            with open(fn, "rb") as fp:
-                data = loadfunc(fp)
-                data = self._data_transform(data)
-                yield (fn, data)
+        with open(path, "rb") as fp:
+            return loadfunc(fp)