+++ src/check_jsonschema/schema_loader/errors.py
@@ -1,7 +1,3 @@
-class BadFileTypeError(ValueError):
-    pass
-
-
 class SchemaParseError(ValueError):
     pass